[onert/test] Split GenModelTests.cc
- Header file for GenModelTest class
- Divide test for each operation
